Tesla - switched FSD 14 detection to regex patterns
