+++ opendbc/car/tesla/carstate.py
@@ -124,13 +124,13 @@ class CarState(CarStateBase):
       # 1. If in Autosteer or FSD, already caught by invalidLkasSetting
       # 2. If in TACC and DAS ever sends ANGLE_CONTROL (1), we can infer it's trying to do LKAS on FSD 14+
       angle_control = cp_ap_party.vl["DAS_steeringControl"]["DAS_steeringControlType"] == 1  # ANGLE_CONTROL
-      if not ret.invalidLkasSetting and angle_control and not self.CP.flags & TeslaFlags.FSD_14 and not self.CP.flags & TeslaFlags.HW3:
+      if not ret.invalidLkasSetting and angle_control and not self.CP.flags & TeslaFlags.FSD_14 and not self.CP.flags & TeslaFlags.NON_FSD_14_SELFDRIVE:
         self.suspected_fsd14 = True
 
       if self.suspected_fsd14:
         ret.invalidLkasSetting = True
         if not self.fsd14_error_logged:
-          carlog.error("FSD 14 detected, but FW not in FSD_14_FW or HW3_FW set")
+          carlog.error("possible FSD 14 detected, but FW did not match FSD 14 rules.")
           self.fsd14_error_logged = True
 
     # Buttons # ToDo: add Gap adjust button
