+++ opendbc/car/tesla/interface.py
@@ -2,7 +2,7 @@ from opendbc.car import Bus, get_safety_config, structs
 from opendbc.car.interfaces import CarInterfaceBase
 from opendbc.car.tesla.carcontroller import CarController
 from opendbc.car.tesla.carstate import CarState
-from opendbc.car.tesla.values import TeslaSafetyFlags, TeslaFlags, CANBUS, CAR, DBC, FSD_14_FW, HW3_FW, Ecu
+from opendbc.car.tesla.values import TeslaSafetyFlags, TeslaFlags, CANBUS, CAR, DBC, FSD_14_FW_RULES, NON_FSD_14_SELFDRIVE_FW_RULES, Ecu, match_rules
 from opendbc.car.tesla.radar_interface import RadarInterface, RADAR_START_ADDR
 
 
@@ -43,14 +43,15 @@ class CarInterface(CarInterfaceBase):
       ret.vEgoStarting = 0.1
       ret.stoppingDecelRate = 0.3
 
-    fsd_14 = any(fw.ecu == Ecu.eps and fw.fwVersion in FSD_14_FW.get(candidate, []) for fw in car_fw)
+    fsd_14 = any(fw.ecu == Ecu.eps and match_rules(FSD_14_FW_RULES.get(candidate, []), fw.fwVersion) for fw in car_fw)
     if fsd_14:
       ret.flags |= TeslaFlags.FSD_14.value
       ret.safetyConfigs[0].safetyParam |= TeslaSafetyFlags.FSD_14.value
 
-    hw3 = any(fw.ecu == Ecu.eps and fw.fwVersion in HW3_FW.get(candidate, []) for fw in car_fw)
-    if hw3:
-      ret.flags |= TeslaFlags.HW3.value
+    non_fsd_14_selfdrive = any(fw.ecu == Ecu.eps and match_rules(NON_FSD_14_SELFDRIVE_FW_RULES.get(candidate, []), fw.fwVersion) for fw in car_fw)
+    if non_fsd_14_selfdrive:
+      ret.flags |= TeslaFlags.NON_FSD_14_SELFDRIVE.value
+      ret.safetyConfigs[0].safetyParam |= TeslaSafetyFlags.NON_FSD_14_SELFDRIVE.value
 
     ret.dashcamOnly = candidate in (CAR.TESLA_MODEL_X,)  # dashcam only, pending find invalidLkasSetting signal
